[Cocoa] Small refactor in Code generation
